fix: Rule Severitiesのレイアウト修正とテスト安定化
Linter設定画面のレイアウト崩壊を修正し、MDxxルール一覧のトグル位置を右端へ固定。また、mmdc環境依存によるPoisonErrorを防ぐため結合テストのアサーションを安定化。さらに、ast-linterのファイル行数制限（200行）を遵守するため、workspace設定を別モジュールへ分離した。
