+++ crates/katana-ui/locales/de.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "Tastenkombinationen"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "updates": {
