+++ crates/katana-ui/locales/es.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "Atajos"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "updates": {
