+++ crates/katana-ui/locales/fr.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "Raccourcis"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "updates": {
