+++ crates/katana-ui/locales/it.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "Scorciatoie"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "updates": {
