+++ crates/katana-ui/locales/ja.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "ショートカット"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "shortcuts": {
