+++ crates/katana-ui/locales/ko.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "단축키"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "updates": {
