+++ crates/katana-ui/locales/zh-TW.json
@@ -353,6 +353,10 @@
       {
         "key": "shortcuts",
         "name": "快捷鍵"
+      },
+      {
+        "key": "linter",
+        "name": "Linter"
       }
     ],
     "updates": {
